+++ build.sh
@@ -136,7 +136,7 @@ main() {
 
     # Get version from version.json
     local version=$(jq -r '.INSTALLER_VERSION' version.json)
-    log_info "Building version: $version"
+    log_info "Building version: [$version]"
 
     # Cleanup previous builds
     log_info "Cleaning up previous builds..."
@@ -170,11 +170,13 @@ main() {
             sed -i "" "s|${INSTALLER_URL}|${INSTALLER_URL}/snapshot|g" build/install-boxlang.sh
             sed -i "" "s|${INSTALLER_URL}|${INSTALLER_URL}/snapshot|g" build/install-boxlang.ps1
             sed -i "" "s|${INSTALLER_URL}|${INSTALLER_URL}/snapshot|g" build/install-bvm.sh
+			sed -i "" "s|@build.version@|${version}|g" build/bvm.sh
         else
             # Linux doesn't use empty string after -i
             sed -i "s|${INSTALLER_URL}|${INSTALLER_URL}/snapshot|g" build/install-boxlang.sh
             sed -i "s|${INSTALLER_URL}|${INSTALLER_URL}/snapshot|g" build/install-boxlang.ps1
             sed -i "s|${INSTALLER_URL}|${INSTALLER_URL}/snapshot|g" build/install-bvm.sh
+			sed -i "s|@build.version@|${version}|g" build/bvm.sh
         fi
     else
 		log_info "Standard build detected, using default installer URL..."